casting issue

+++ libs/core/include/core/fixed_point/fixed_point.hpp
@@ -208,7 +208,7 @@ struct BaseFixedpointType
 template <std::uint16_t I, std::uint16_t F>
 class FixedPoint : public BaseFixedpointType
 {
-  static_assert(TypeFromSize<I + F>::is_valid, "invalid combination of sizes");
+  static_assert(TypeFromSize<static_cast<int>(I + F)>::is_valid, "invalid combination of sizes");
 
 public:
   enum
